KnowBargain(Backend)/requirements.txt
flask
flask_sqlalchemy
flask_cors
pyjwt
cloudinary
